+++ src/bfabric_scripts/cli/cli_executable.py
@@ -1,0 +1,6 @@
+import cyclopts
+
+from bfabric_scripts.cli.executable.inspect import inspect_executable
+
+app = cyclopts.App()
+app.command(inspect_executable, name="inspect")